Sort + Spect + Pagination en ClientController
Quedan marcas (todo) en el codigo, principalmente con excepciones que no muestran contenido

+++ tpFinal/src/main/java/com/utn/tpFinal/service/ClientService.java
@@ -1,8 +1,12 @@
 package com.utn.tpFinal.service;
 
+import com.utn.tpFinal.exception.ClientNotExists;
+import com.utn.tpFinal.exception.NoContentException;
+import com.utn.tpFinal.exception.ResidenceNotExists;
 import com.utn.tpFinal.model.Client;
 import com.utn.tpFinal.model.Residence;
 import com.utn.tpFinal.model.dto.ClientDto;
+import com.utn.tpFinal.model.dto.ResidenceDto;
 import com.utn.tpFinal.model.proyection.Top10Clients;
 import com.utn.tpFinal.repository.ClientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,12 +14,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.Sort.Order;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpClientErrorException;
-
 
 import java.util.Date;
 import java.util.List;
@@ -25,8 +29,6 @@ import static java.util.Objects.isNull;
 @Service
 public class ClientService {
 
-    private static final String CLIENT_PATH ="client";
-
     private ClientRepository clientRepository;
     private ResidenceService residenceService;
     private BillService billService;
@@ -35,7 +37,11 @@ public class ClientService {
     public ClientService(ClientRepository clientRepository, ResidenceService residenceService) {
         this.clientRepository = clientRepository;
         this.residenceService = residenceService;
+    }
 
+    public Client getClientById(Integer id) throws ClientNotExists {
+        return clientRepository.findById(id)
+                .orElseThrow(() -> new ClientNotExists(this.getClass().getSimpleName(), "getClientById"));
     }
 
 //-------------------------------------------->> M E T O D O S <<--------------------------------------------
@@ -44,55 +50,56 @@ public class ClientService {
         return clientRepository.save(client);
     }
 
-    public Page<Client> getAll(String name, Integer pageNumber, Integer pageSize, String sortBy) {
-        Pageable pageable = PageRequest.of(pageNumber, pageSize, Sort.by(sortBy));
-        Page<Client> pagedResult;
+    public Page<ClientDto> getAll(Specification<Client> clientSpecification, Integer page, Integer size, List<Order>orders) throws NoContentException {
+        Pageable pageable = PageRequest.of(page, size, Sort.by(orders));
+        Page<Client>clients = clientRepository.findAll(clientSpecification,pageable);
 
-        if(isNull(name))
-            pagedResult = clientRepository.findAll(pageable);
-        else
-            pagedResult = clientRepository.findByName(name,pageable);
+        if(clients.isEmpty())
+            throw new NoContentException(this.getClass().getSimpleName(), "getAll");
 
-        return pagedResult;
+        Page<ClientDto> dtoClients = clients.map(c -> ClientDto.fromWithOutResidences(c));
+
+        return dtoClients;
     }
 
-
-
-    public Client getClientById(Integer id) {
-        return clientRepository.findById(id)
-                .orElseThrow(() -> new HttpClientErrorException(HttpStatus.NOT_FOUND));
-    }
-
-    public Boolean clientExists(Integer id) {
-        return clientRepository.existsById(id);
-    }
-
-    public void addResidenceToClient(Integer idClient, Integer idResidence) {
-        Client c = getClientById(idClient);
-        Residence r = residenceService.getResidenceById(idResidence);
+    public void addResidenceToClient(Integer idClient, Integer idResidence) throws ClientNotExists, ResidenceNotExists {
+        Client c = getClientById(idClient); //devuelve exception
+        Residence r = residenceService.getResidenceById(idResidence); //devuelve exception
         residenceService.addClientToResidence(c, r);
         c.getResidencesList().add(r);
         clientRepository.save(c);
     }
 
-    public List<Residence> getClientResidences(Integer idClient) {
-        Client c = getClientById(idClient);
-        return c.getResidencesList();
+    public Page<ResidenceDto> getClientResidences(Integer idClient, Integer page, Integer size, List<Order>orders) throws NoContentException, ClientNotExists {
+        if(!clientRepository.existsById(idClient))
+            throw new ClientNotExists(this.getClass().getSimpleName(), "getClientResidences");
+
+        Pageable pageable = PageRequest.of(page, size, Sort.by(orders));
+        Page<Residence> residences = residenceService.getResidenceByClientId(idClient, pageable);
+
+        if(residences.isEmpty()) //todo no muestra el contenido de la exception
+            throw new NoContentException(this.getClass().getSimpleName(), "getClientResidences");
+
+        Page<ResidenceDto> residencesDto = residences.map(r-> ResidenceDto.from(r));
+
+        return residencesDto;
     }
 
-    public void deleteClientById(Integer idClient) {
+    public void deleteClientById(Integer idClient) throws ClientNotExists {
+        if(!clientRepository.existsById(idClient))
+            throw new ClientNotExists(this.getClass().getSimpleName(), "deleteClientById");
+
         clientRepository.deleteById(idClient);
     }
 
-    public List<Top10Clients> getTop10ConsumerByDates(Date from, Date to) {
-        return clientRepository.getTop10Clients(from,to);
+    public List<Top10Clients> getTop10ConsumerByDates(Date from, Date to) throws NoContentException {
+        List<Top10Clients> clientsList = clientRepository.getTop10Clients(from,to);
+        if(clientsList.isEmpty()) //todo no muestra el contenido de la exception
+            throw new NoContentException(this.getClass().getSimpleName(), "getTop10ConsumerByDates");
+        return clientsList;
     }
 
-    public Page<ClientDto> getAll(Specification<Client> clientSpecification, Pageable pageable) {
-        Page<Client>clients = clientRepository.findAll(clientSpecification, pageable);
-        Page<ClientDto> dtoClients = clients.map(c -> ClientDto.fromWithOutResidences(c));
-        return dtoClients;
-    }
+
 
 
 }
